requirement.txt
flask
requests
gevent
Werkzeug
python_magic
tensorflow
pillow
numpy
